LayoutTests/fast/dom/getElementsByClassName/non-styled-element-expected.txt

Tests that getElementsByClassName() is able to return non-styled elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS collection.length is 1
PASS collection[0] is nonStyledElement
PASS successfullyParsed is true

TEST COMPLETE
